+++ src/app.js
@@ -1,10 +1,14 @@
 import './styles/style.css'
 import cookie from './modules/cookie'
-import core from './modules/core'
+import bongoCat from './modules/bongoCat'
 import gtag from './modules/gtag'
 import {internationalize} from './modules/internationalize'
-$(window).on('load', cookie())
+
+// Init modules
+bongoCat()
 gtag()
-core()
+
+// Subscribe  event's callbacks
+$(window).on('load', cookie())
 $(window).on("languagechange", internationalize)
 $(document).on("DOMContentLoaded", internationalize)
